requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-WTF>=1.2.1
Flask-Login>=0.6.3
Flask-Scss>=0.5
WTForms[email]>=3.1.2
email-validator>=2.2.0
Werkzeug>=3.0.0
python-dotenv>=1.0.1
pyScss>=1.4.0
cryptography>=43.0.0
requests>=2.32.0